+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project.
 - `scope` and `language_code` fields to `SetMyCommands` request class.
 - `DeleteMyCommands` request class.
 
+### Changed
+- `KickChatMember` renamed to `BanChatMember`. The old method name can still be used.
+
+### Deprecated
+- `KickChatMember` is deprecated and will be removed in version 4.
+
 ### Fixed
 - typing for `Message` object.